requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
albumentations==0.5.2
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
ansi2html==1.8.0
antlr4-python3-runtime==4.8
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astropy==4.3.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
braceexpand==0.1.7
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
crcmod==1.7
cryptography==41.0.5
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.13.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
dill==0.3.6
easydict==1.9
einops==0.6.1
elasticdeform==0.5.0
entrypoints==0.4
exceptiongroup==1.1.3
fastjsonschema==2.18.0
filelock==3.12.2
fire==0.5.0
flake8==3.9.2
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
ghp-import==2.1.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.8.0
huggingface-hub==0.16.4
hydra-core==1.1.0
idna==3.4
imagecorruptions==1.1.2
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_packaging==0.12.3
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.11.0
kiwisolver==1.4.5
kornia==0.5.0
libclang==16.0.6
llvmlite==0.39.1
lyft-dataset-sdk==0.0.8
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.1
mmcv==2.1.0
mmdet==3.2.0
mmdet3d==1.3.0
mmengine==0.9.0
model-index==0.1.11
modelscope==1.9.4
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
natsort==8.4.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.7.0
nest-asyncio==1.5.7
networkx==2.6.3
notebook==6.5.5
notebook_shim==0.2.3
numba==0.56.4
numpy==1.21.6
nuscenes-devkit==1.1.10
oauthlib==3.2.2
omegaconf==2.1.2
open3d==0.17.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.2.0
plyfile==0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.19.0
pyerfa==2.0.0.3
pyflakes==2.3.1
Pygments==2.16.1
PyMieScatt==1.8.1.1
pyparsing==3.1.1
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.4.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.2.9
pytz==2023.3.post1
PyWavelets==1.3.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.7.3
Send2Trash==1.8.2
Shapely==1.8.5
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchmetrics==0.2.0
torchvision==0.9.0+cu111
tornado==6.2
tqdm==4.65.2
traitlets==5.9.0
trimesh==4.0.1
typed-ast==1.5.5
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.73
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wincertstore==0.2
wldhx.yadisk-direct==0.0.6
wrapt==1.15.0
xxhash==3.4.1
y-py==0.6.0
yapf==0.40.2
yarl==1.9.2
ypy-websocket==0.8.4
zipp==3.15.0
